+++ src/Routes/RouteLayout.js
@@ -22,7 +22,8 @@ import Check_Email from '../Pages/Check_Email/Check_Email'
 import Verification from '../Pages/Verification/Verification'
 import CheckOut from '../Pages/CheckOut/CheckOut'
 import Product_List_Page from '../Pages/Product_List_Page/Product_List_Page'
-import Contact_Details from '../Pages/Contact_Details/Contact_Details.jsx/Contact_Details'
+import Contact_Details from '../Pages/Contact_Details/Contact_Details_Right/Contact_Details'
+
 
 
 const RouteLayout = () => {
